+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -6,6 +6,7 @@ package mozilla.components.concept.engine
 
 import android.graphics.Bitmap
 import androidx.annotation.CallSuper
+import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.engine.media.RecordingDevice
 import mozilla.components.concept.engine.permission.PermissionRequest
@@ -14,7 +15,6 @@ import mozilla.components.concept.engine.window.WindowRequest
 import mozilla.components.support.base.observer.Observable
 import mozilla.components.support.base.observer.ObserverRegistry
 import java.lang.UnsupportedOperationException
-import org.json.JSONObject
 
 /**
  * Class representing a single engine session.
@@ -51,7 +51,7 @@ abstract class EngineSession(
         fun onCloseWindowRequest(windowRequest: WindowRequest) = Unit
         fun onMediaAdded(media: Media) = Unit
         fun onMediaRemoved(media: Media) = Unit
-        fun onWebAppManifestLoaded(manifest: JSONObject) = Unit
+        fun onWebAppManifestLoaded(manifest: WebAppManifest) = Unit
         fun onCrashStateChange(crashed: Boolean) = Unit
         fun onRecordingStateChanged(devices: List<RecordingDevice>) = Unit